+++ SGPP/database/migrations/2019_03_10_160028_create_entrada_seguimientos_table.php
@@ -18,7 +18,7 @@ class CreateEntradaSeguimientosTable extends Migration
             $table->string('actividad');
             $table->string('observacion');
             $table->string('urlEvidencias');
-            $table->integer('horasRealizadas');
+            $table->decimal('horasRealizadas');
             $table->boolean('validado');
             $table->boolean('comprobado');
             $table->integer('asignacion_id')->unsigned();
